Seperate compnent concerns. Add useEffect to make API call for token
