+++ src/Header.jsx
@@ -1,0 +1,7 @@
+export default function Header() {
+    return (
+        <header className="header-container">
+    <h1>Ja<span>mmm</span>ing</h1>
+    </header>
+    );
+}
