+++ src/Playlists.jsx
@@ -1,0 +1,8 @@
+export default function Playlists(){
+    return (
+        <article className="playlist-container">
+            <h2 className="header">Playlists</h2>
+            <button>Save Playlist</button>
+        </article>
+    );
+}
