+++ src/Results.jsx
@@ -1,0 +1,7 @@
+export default function Results(){
+    return (
+        <article className="results-container">
+            <h2 className="header">Results</h2>
+        </article>
+    )
+}
