+++ src/Search.jsx
@@ -1,0 +1,9 @@
+export default function Search(){
+    return (
+        <section>
+            <input type="text" defaultValue="Artist, Song, Genre, etc." />
+        <br/>
+            <button>Search</button>
+      </section>
+    );
+}
